fixme: split daphnia_ruler into helpers. problem with arguments
